Warning cleanup
Builds no longer use -pedantic
B2 builds now use -Wextra
Fixed various warnings

close #186

+++ include/boost/mysql/impl/resultset.ipp
@@ -21,7 +21,7 @@ void boost::mysql::resultset::assign(resultset_view v)
         rws_ = v.rows();
         affected_rows_ = v.affected_rows();
         last_insert_id_ = v.last_insert_id();
-        warnings_ = v.warning_count();
+        warnings_ = static_cast<std::uint16_t>(v.warning_count());
         info_.assign(v.info().begin(), v.info().end());
         is_out_params_ = v.is_out_params();
     }
